+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 v0.3.0 (2014-xx-xx)
 ===============================================================================
 
+- Switched to PSR-4 for autoloading.
+
 - Switched to react/event-loop 0.4 and predis/predis 1.0. This change breaks the
   compatibility with previous versions of Predis\Async due to the changes needed
   to adapt to the new (and stable) API of Predis. Support for PHP 5.3 has been
   dropped since newer versions of React require PHP >= 5.4.
 
-- Switched to PSR-4 for autoloading.
+- The phpiredis extension is now optional and by default the client uses a pure
+  PHP protocol serializer / parser provided by the clue/redis-protocol library.
 
 v0.2.2 (2013-09-04)
 ===============================================================================